[APT-482] - implement tests & implementation for NotificationReportDao
created mapping into DTO classes (as helper methods inside Repository)
created queries for all three dao methods
created result mappers for all queries
test using random-beans and java instead of insert statements
(pros -> we cannot hold id's and values twice. in statement & in code e.g. as constants)

+++ backend/src/main/java/com/creatix/domain/dto/notification/reporting/NotificationReportGroupByAccountDto.java
@@ -12,8 +12,8 @@ public class NotificationReportGroupByAccountDto {
 
     private final Long confirmed;
     private final Long resolved;
-    private final Long averageTimeToConfirm;
-    private final Long averageTimeToResolve;
+    private final Double averageTimeToConfirm;
+    private final Double averageTimeToResolve;
 
     @Setter
     private NotificationReportAccountDto account;
